chore: re-creating account.go for controller side for consitency
