+++ modules/apps/27-interchain-accounts/controller/keeper/keeper.go
@@ -7,13 +7,11 @@ import (
 	baseapp "github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	capabilitykeeper "github.com/cosmos/cosmos-sdk/x/capability/keeper"
 	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
 	"github.com/tendermint/tendermint/libs/log"
 
 	"github.com/cosmos/ibc-go/v2/modules/apps/27-interchain-accounts/types"
-	channeltypes "github.com/cosmos/ibc-go/v2/modules/core/04-channel/types"
 	host "github.com/cosmos/ibc-go/v2/modules/core/24-host"
 )
 
@@ -53,36 +51,6 @@ func NewKeeper(
 // Logger returns the application logger, scoped to the associated module
 func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 	return ctx.Logger().With("module", fmt.Sprintf("x/%s-%s", host.ModuleName, types.ModuleName))
-}
-
-// InitInterchainAccount is the entry point to registering an interchain account.
-// It generates a new port identifier using the owner address, connection identifier,
-// and counterparty connection identifier. It will bind to the port identifier and
-// call 04-channel 'ChanOpenInit'. An error is returned if the port identifier is
-// already in use. Gaining access to interchain accounts whose channels have closed
-// cannot be done with this function. A regular MsgChanOpenInit must be used.
-func (k Keeper) InitInterchainAccount(ctx sdk.Context, connectionID, counterpartyConnectionID, owner string) error {
-	portID, err := types.GeneratePortID(owner, connectionID, counterpartyConnectionID)
-	if err != nil {
-		return err
-	}
-
-	if k.portKeeper.IsBound(ctx, portID) {
-		return sdkerrors.Wrap(types.ErrPortAlreadyBound, portID)
-	}
-
-	cap := k.BindPort(ctx, portID)
-	if err := k.ClaimCapability(ctx, cap, host.PortPath(portID)); err != nil {
-		return sdkerrors.Wrap(err, "unable to bind to newly generated portID")
-	}
-
-	msg := channeltypes.NewMsgChannelOpenInit(portID, types.VersionPrefix, channeltypes.ORDERED, []string{connectionID}, types.PortID, types.ModuleName)
-	handler := k.msgRouter.Handler(msg)
-	if _, err := handler(ctx, msg); err != nil {
-		return err
-	}
-
-	return nil
 }
 
 // GetAllPorts returns all ports to which the interchain accounts controller module is bound. Used in ExportGenesis
